requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
elevenlabs==1.3.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyNaCl==1.5.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
